Revert "Merge branch 'code/anastasia-e/stand-behaviour' into MVP/august"
This reverts commit 109382d8a5d0e92ef211a74fd68076a34c946a84, reversing
changes made to decd1807115f49ad92f06c3b5f7fa038da726ac2.
