--- Rescues/Assets/Resources/Data/StandItems.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 236d5357b918c4145af08489adfe9012
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
