--- Rescues/Assets/Resources/Data/StandItems/TestData1.asset
@@ -1,15 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ff1e0270b226e2c4796fd8eb33bb7a1d, type: 3}
-  m_Name: TestData1
-  m_EditorClassIdentifier: 
-  Sprite: {fileID: 21300000, guid: 490641638817ea54a89dfc596e780979, type: 3}
