--- Rescues/Assets/Resources/Data/StandItems/TestData1.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e35afb0f81f747341860d09c3aea085b
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
