--- Rescues/Assets/Resources/Data/StandItems/TestData2.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b48a1287a5f4b0a49897de4a612c3606
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
