--- Rescues/Assets/Resources/Prefabs/TestPrefabs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 73b3896b6a2a6a24bae2579685d5fe58
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
