--- Rescues/Assets/Resources/Prefabs/TestPrefabs/Stand.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 329d665178bc74b429ab0d359358f3a6
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
