--- Rescues/Assets/Resources/Prefabs/UI/StandWindow.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 42abd26bd69c1044d97b44a89d1e93dc
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
