+++ Rescues/Assets/Scripts/Controllers/GameControllers/MainControllers.cs
@@ -14,7 +14,6 @@
             Add(new DoorTeleporterController(context, services));
             Add(new MainPuzzleController(context, services));
             Add(new HidingPlaceController(context, services));
-            Add(new StandController(context, services));
             Add(new EnemyVisionController(context, services));
             Add(new EnemyMoveController(context, services));
             Add(new InputController(context, services));
