+++ Rescues/Assets/Scripts/Controllers/HidingPlaceController.cs.meta
