--- Rescues/Assets/Scripts/Controllers/InteractableObject/StandController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 61359077d7ecd4448841c4e542f61f21
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
