--- Rescues/Assets/Scripts/Data/StandItemData.cs
@@ -1,13 +1,0 @@
-﻿using UnityEngine;
-using UnityEngine.UI;
-
-
-[CreateAssetMenu(fileName = "StandItemData", menuName = "Data/StandItemData")]
-public class StandItemData : ScriptableObject
-{
-    #region Fields
-
-    public Sprite Sprite;  
-
-    #endregion
-}
