--- Rescues/Assets/Scripts/Data/StandItemData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ff1e0270b226e2c4796fd8eb33bb7a1d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
