+++ Rescues/Assets/Scripts/Enum/InteractableObjectType.cs
@@ -8,6 +8,5 @@
         HidingPlace = 3,
         Trap = 4,
         Puzzle = 5,
-        Stand  = 6,
     }
 }
