+++ Rescues/Assets/Scripts/Model/Behaviour/HidingPlaceBehaviour.cs
@@ -8,7 +8,6 @@ namespace Rescues
         #region Fields
         
         public HidingPlaceData HidingPlaceData;
-        public SpriteRenderer HidedSprite;
 
         #endregion
     }
