--- Rescues/Assets/Scripts/Model/Behaviour/StandBehaviour.cs
@@ -1,14 +1,0 @@
-﻿using UnityEngine;
-
-
-namespace Rescues
-{
-    public sealed class StandBehaviour: InteractableObjectBehavior
-    {
-        #region Fields
-
-        public GameObject StandWindow;
-        
-        #endregion
-    }
-}
