--- Rescues/Assets/Scripts/Model/Stand.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 038b744bfa4b9d548bfef97c9bf260b4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
