--- Rescues/Assets/Scripts/Model/Stand/StandItem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8afede7f54c952f4ca8005770a3b5739
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
