--- Rescues/Assets/Scripts/Model/Stand/StandUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b3eab52be1f09b543b28709e4e444229
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
